Split sings, arrays, objects and colors into modules
