+++ lib/anime.umd.js
@@ -13,6 +13,8 @@
 
   var version = "3.3.0";
 
+  // Default animation parameters
+
   const defaultInstanceSettings = {
     update: null,
     begin: null,
@@ -35,6 +37,8 @@
     easing: 'easeOutElastic(1, .5)',
     round: 0
   };
+
+  // Transforms
 
   const validTransforms = [
     'translateX',
@@ -60,6 +64,8 @@
     CSS: {},
     springs: {}
   };
+
+  // Utility functions
 
   function minMax(val, min, max) {
     return Math.min(Math.max(val, min), max);
@@ -90,6 +96,71 @@
     col: a => (is.hex(a) || is.rgb(a) || is.hsl(a)),
     key: a => !defaultInstanceSettings.hasOwnProperty(a) && !defaultTweenSettings.hasOwnProperty(a) && a !== 'targets' && a !== 'keyframes',
   };
+
+  // Strings
+
+  function selectString(str) {
+    try {
+      let nodes = document.querySelectorAll(str);
+      return nodes;
+    } catch(e) {
+      return;
+    }
+  }
+
+  // Arrays
+
+  function filterArray(arr, callback) {
+    const len = arr.length;
+    const thisArg = arguments.length >= 2 ? arguments[1] : void 0;
+    const result = [];
+    for (let i = 0; i < len; i++) {
+      if (i in arr) {
+        const val = arr[i];
+        if (callback.call(thisArg, val, i, arr)) {
+          result.push(val);
+        }
+      }
+    }
+    return result;
+  }
+
+  function flattenArray(arr) {
+    return arr.reduce((a, b) => a.concat(is.arr(b) ? flattenArray(b) : b), []);
+  }
+
+  function toArray(o) {
+    if (is.arr(o)) return o;
+    if (is.str(o)) o = selectString(o) || o;
+    if (o instanceof NodeList || o instanceof HTMLCollection) return [].slice.call(o);
+    return [o];
+  }
+
+  function arrayContains(arr, val) {
+    return arr.some(a => a === val);
+  }
+
+  // Objects
+
+  function cloneObject(o) {
+    const clone = {};
+    for (let p in o) clone[p] = o[p];
+    return clone;
+  }
+
+  function replaceObjectProps(o1, o2) {
+    const o = cloneObject(o1);
+    for (let p in o1) o[p] = o2.hasOwnProperty(p) ? o2[p] : o1[p];
+    return o;
+  }
+
+  function mergeObjects(o1, o2) {
+    const o = cloneObject(o1);
+    for (let p in o2) o[p] = is.und(o1[p]) ? o2[p] : o1[p];
+    return o;
+  }
+
+  // Easings
 
   function parseEasingParameters(string) {
     const match = /\(([^)]+)\)/.exec(string);
@@ -286,69 +357,6 @@
     }
   }
 
-  // Strings
-
-  function selectString(str) {
-    try {
-      let nodes = document.querySelectorAll(str);
-      return nodes;
-    } catch(e) {
-      return;
-    }
-  }
-
-  // Arrays
-
-  function filterArray(arr, callback) {
-    const len = arr.length;
-    const thisArg = arguments.length >= 2 ? arguments[1] : void 0;
-    const result = [];
-    for (let i = 0; i < len; i++) {
-      if (i in arr) {
-        const val = arr[i];
-        if (callback.call(thisArg, val, i, arr)) {
-          result.push(val);
-        }
-      }
-    }
-    return result;
-  }
-
-  function flattenArray(arr) {
-    return arr.reduce((a, b) => a.concat(is.arr(b) ? flattenArray(b) : b), []);
-  }
-
-  function toArray(o) {
-    if (is.arr(o)) return o;
-    if (is.str(o)) o = selectString(o) || o;
-    if (o instanceof NodeList || o instanceof HTMLCollection) return [].slice.call(o);
-    return [o];
-  }
-
-  function arrayContains(arr, val) {
-    return arr.some(a => a === val);
-  }
-
-  // Objects
-
-  function cloneObject(o) {
-    const clone = {};
-    for (let p in o) clone[p] = o[p];
-    return clone;
-  }
-
-  function replaceObjectProps(o1, o2) {
-    const o = cloneObject(o1);
-    for (let p in o1) o[p] = o2.hasOwnProperty(p) ? o2[p] : o1[p];
-    return o;
-  }
-
-  function mergeObjects(o1, o2) {
-    const o = cloneObject(o1);
-    for (let p in o2) o[p] = is.und(o1[p]) ? o2[p] : o1[p];
-    return o;
-  }
-
   // Colors
 
   function rgbToRgba(rgbValue) {
@@ -366,20 +374,21 @@
     return `rgba(${r},${g},${b},1)`;
   }
 
+  function hue2rgb(p, q, t) {
+    if (t < 0) t += 1;
+    if (t > 1) t -= 1;
+    if (t < 1/6) return p + (q - p) * 6 * t;
+    if (t < 1/2) return q;
+    if (t < 2/3) return p + (q - p) * (2/3 - t) * 6;
+    return p;
+  }
+
   function hslToRgba(hslValue) {
     const hsl = /hsl\((\d+),\s*([\d.]+)%,\s*([\d.]+)%\)/g.exec(hslValue) || /hsla\((\d+),\s*([\d.]+)%,\s*([\d.]+)%,\s*([\d.]+)\)/g.exec(hslValue);
     const h = parseInt(hsl[1], 10) / 360;
     const s = parseInt(hsl[2], 10) / 100;
     const l = parseInt(hsl[3], 10) / 100;
     const a = hsl[4] || 1;
-    function hue2rgb(p, q, t) {
-      if (t < 0) t += 1;
-      if (t > 1) t -= 1;
-      if (t < 1/6) return p + (q - p) * 6 * t;
-      if (t < 1/2) return q;
-      if (t < 2/3) return p + (q - p) * (2/3 - t) * 6;
-      return p;
-    }
     let r, g, b;
     if (s == 0) {
       r = g = b = l;
@@ -393,10 +402,10 @@
     return `rgba(${r * 255},${g * 255},${b * 255},${a})`;
   }
 
-  function colorToRgb(val) {
-    if (is.rgb(val)) return rgbToRgba(val);
-    if (is.hex(val)) return hexToRgba(val);
-    if (is.hsl(val)) return hslToRgba(val);
+  function normalizeColorToRgba(colorValue) {
+    if (is.rgb(colorValue)) return rgbToRgba(colorValue);
+    if (is.hex(colorValue)) return hexToRgba(colorValue);
+    if (is.hsl(colorValue)) return hslToRgba(colorValue);
   }
 
   // Units
@@ -497,7 +506,7 @@
   }
 
   function validateValue(val, unit) {
-    if (is.col(val)) return colorToRgb(val);
+    if (is.col(val)) return normalizeColorToRgba(val);
     if (/\s/g.test(val)) return val;
     const originalUnit = getUnit(val);
     const unitLess = originalUnit ? val.substr(0, val.length - originalUnit.length) : val;
