+++ src/easings.js
@@ -6,6 +6,8 @@ import {
 
 import cache from './cache.js';
 
+// Easings
+
 function parseEasingParameters(string) {
   const match = /\(([^)]+)\)/.exec(string);
   return match ? match[1].split(',').map(p => parseFloat(p)) : [];
