+++ src/helpers.js
@@ -3,6 +3,8 @@ import {
   defaultTweenSettings
 } from './consts.js';
 
+// Utility functions
+
 function minMax(val, min, max) {
   return Math.min(Math.max(val, min), max);
 }
@@ -33,9 +35,80 @@ const is = {
   key: a => !defaultInstanceSettings.hasOwnProperty(a) && !defaultTweenSettings.hasOwnProperty(a) && a !== 'targets' && a !== 'keyframes',
 }
 
+// Strings
+
+function selectString(str) {
+  try {
+    let nodes = document.querySelectorAll(str);
+    return nodes;
+  } catch(e) {
+    return;
+  }
+}
+
+// Arrays
+
+function filterArray(arr, callback) {
+  const len = arr.length;
+  const thisArg = arguments.length >= 2 ? arguments[1] : void 0;
+  const result = [];
+  for (let i = 0; i < len; i++) {
+    if (i in arr) {
+      const val = arr[i];
+      if (callback.call(thisArg, val, i, arr)) {
+        result.push(val);
+      }
+    }
+  }
+  return result;
+}
+
+function flattenArray(arr) {
+  return arr.reduce((a, b) => a.concat(is.arr(b) ? flattenArray(b) : b), []);
+}
+
+function toArray(o) {
+  if (is.arr(o)) return o;
+  if (is.str(o)) o = selectString(o) || o;
+  if (o instanceof NodeList || o instanceof HTMLCollection) return [].slice.call(o);
+  return [o];
+}
+
+function arrayContains(arr, val) {
+  return arr.some(a => a === val);
+}
+
+// Objects
+
+function cloneObject(o) {
+  const clone = {};
+  for (let p in o) clone[p] = o[p];
+  return clone;
+}
+
+function replaceObjectProps(o1, o2) {
+  const o = cloneObject(o1);
+  for (let p in o1) o[p] = o2.hasOwnProperty(p) ? o2[p] : o1[p];
+  return o;
+}
+
+function mergeObjects(o1, o2) {
+  const o = cloneObject(o1);
+  for (let p in o2) o[p] = is.und(o1[p]) ? o2[p] : o1[p];
+  return o;
+}
+
 export {
   minMax,
   stringContains,
   applyArguments,
-  is
+  is,
+  selectString,
+  filterArray,
+  flattenArray,
+  toArray,
+  arrayContains,
+  cloneObject,
+  replaceObjectProps,
+  mergeObjects
 }
